# ajk/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Emergency Alert API",
        "description": "Emergency Alert API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/admin/users/{login}": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUserDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminUserDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "updateUser_1",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUserDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUserDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "public-user-resource"
                ],
                "operationId": "getAllPublicUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminUserDTO": {
                "required": [
                    "login"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "maxLength": 50,
                        "minLength": 1,
                        "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                        "type": "string"
                    },
                    "firstName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "lastName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 254,
                        "minLength": 5,
                        "type": "string"
                    },
                    "imageUrl": {
                        "maxLength": 256,
                        "minLength": 0,
                        "type": "string"
                    },
                    "activated": {
                        "type": "boolean"
                    },
                    "langKey": {
                        "maxLength": 10,
                        "minLength": 2,
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "authorities": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "User": {
                "required": [
                    "activated",
                    "login"
                ],
                "type": "object",
                "properties": {
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "maxLength": 50,
                        "minLength": 1,
                        "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                        "type": "string"
                    },
                    "firstName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "lastName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 254,
                        "minLength": 5,
                        "type": "string"
                    },
                    "activated": {
                        "type": "boolean"
                    },
                    "langKey": {
                        "maxLength": 10,
                        "minLength": 2,
                        "type": "string"
                    },
                    "imageUrl": {
                        "maxLength": 256,
                        "minLength": 0,
                        "type": "string"
                    },
                    "resetDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "UserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "type": "string"
                    }
                }
            }
        }
    }
}